scripts/requirements-skills.txt
redis>=4.5.1
pyyaml>=6.0
python-dotenv>=1.0.0
